requirements.txt
python-telegram-bot>=21.0
openai>=1.12.0
APScheduler>=3.10.4
SQLAlchemy>=2.0.25
asyncpg>=0.29.0
aiosqlite>=0.19.0
alembic>=1.13.1
python-dotenv>=1.0.1
httpx>=0.26.0
matplotlib>=3.8.0
Pillow>=10.2.0
pydantic>=2.6.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
